Adding EVAL/EVALSHA support
This is a hack-y approach and really just exists in case we need EVAL
before we can work out a good solution in the main branch.

The issue is that we're doing a somewhat ad hoc approach to handling
EVAL and EVALSHA rather than working out a good approach to handle other
vector commands that don't need to be split.

This approach assumes that the first key provided hashes to the server
where all the other keys are present.  If you aren't using hash_tags,
you will most likely find that any data written by your Lua script will
not be retrievable.